Make simulation of Monty Hall problem
Simulating the Monty Hall game, where a game host presents 3 doors to a
player. Behind one of the doors a price is to be found, while the other
doors will cover only an empty slot. The player selects one door. Before opening the door, the game
host opens one of the two other doors to show an empty slot. The player
is then asked if he would like to switch from the selected door to the
other, still closed, door.

The theory is that the player will have an improved chance of winning
the price, if he switches door. This simulator should reach the same
conclusion.